+++ afat/tasks.py
@@ -19,6 +19,7 @@ from allianceauth.services.tasks import QueueOnce
 from esi.models import Token
 
 # Alliance Auth (External Libs)
+from app_utils.esi import fetch_esi_status
 from app_utils.logging import LoggerAddTag
 
 # Alliance Auth AFAT
@@ -210,6 +211,12 @@ def update_esi_fatlinks() -> None:
     :return:
     :rtype:
     """
+
+    # Abort if ESI seems to be offline or above error limit
+    if not fetch_esi_status().is_ok:
+        logger.warning("ESI doesn't seem to be available at this time. Aborting.")
+
+        return
 
     required_scopes = ["esi-fleets.read_fleet.v1"]